+++ tests/common/mod.rs
@@ -1,3 +1,5 @@
+pub mod ws_client;
+
 use ajj::{HandlerCtx, Router};
 use serde_json::{json, Value};
 use std::time::Duration;
